docker/consumer/requirements.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
aioredis